+++ modules/core/src/lib/utils/url-utils.ts
@@ -12,3 +12,12 @@ export function extractQueryString(url): string {
 export function stripQueryString(url): string {
   return url.replace(QUERY_STRING_PATTERN, '');
 }
+
+export function shortenUrlForDisplay(url: string): string {
+  if (url.length < 50) {
+    return url;
+  }
+  const urlEnd = url.slice(url.length - 15);
+  const urlStart = url.substr(0, 32);
+  return `${urlStart}...${urlEnd}`;
+}
