+++ modules/core/test/index.ts
@@ -10,6 +10,7 @@ import './iterators/make-stream.spec';
 import './lib/utils/mime-type-utils.spec';
 import './lib/utils/resource-utils.spec';
 import './lib/utils/response-utils.spec';
+import './lib/utils/url-utils.spec';
 
 import './lib/loader-utils/loggers.spec';
 import './lib/loader-utils/option-utils.spec';
